mearge files for alert
