+++ tests/alerts.test.ts
@@ -8,5 +8,15 @@ test("Hendaling Alerts", async ({ page }) => {
     await alert.accept()
   })
   await page.locator("button:has-text('Click Me')").nth(0).click()
-  expect(page.locator("id=confirm-demo")).toContainText("cancle!")
+});
+
+test("Confirm Box", async ({ page }) => {
+  await page.goto('https://www.lambdatest.com/selenium-playground/javascript-alert-box-demo');
+  page.on("dialog",async (alert)=>{
+    const txt=alert.message()
+    console.log(txt)
+    await alert.dismiss()
+  })
+  await page.locator("button:has-text('Click Me')").nth(1).click()
+  expect(page.locator("id=confirm-demo")).toContainText("Cancel!")
 });
